Revert "Merge branch 'alpha' into develop"
This reverts commit 1f30781c32b46edca3569b2d5ba7578ffaaae5fb, reversing
changes made to e5c87b1dd5a3b4e5ae403d2003c5b932fe8e207b.
