+++ .circleci/config.yml
@@ -28,7 +28,7 @@ jobs:
           paths:
             - node_modules
             # add compiled contracts to persisted workspace
-            # (note not currently working due to:
+            # (note not currentlyw orking due to:
             # https://github.com/trufflesuite/truffle/issues/2239#issuecomment-552041859)
             - build
 
@@ -67,14 +67,6 @@ jobs:
             exit $EXIT_CODE
       - run: cat test/publish/test.log
 
-  test-local:
-    <<: *env_defaults
-    steps:
-      - checkout
-      - attach_workspace:
-          at: .
-      - run: npm run test:local
-
   test-kovan:
     <<: *env_defaults
     steps:
@@ -111,9 +103,6 @@ workflows:
           requires:
             - prepare
       - test-publish:
-          requires:
-            - prepare
-      - test-local:
           requires:
             - prepare
       - test-kovan:
