+++ contracts/test-helpers/MockEtherCollateral.sol
@@ -1,3 +1,4 @@
+
 pragma solidity 0.4.25;
 
 import "../SafeDecimalMath.sol";
