+++ publish/deployed/kovan/deployment.json
@@ -20,11 +20,11 @@
 		},
 		"Depot": {
 			"name": "Depot",
-			"address": "0x8F058332CdE72FEb4E062aAB521C5659A36E88b2",
+			"address": "0x73a61f7D7401096ccB4AC883ebeb7bE7605c3955",
 			"source": "Depot",
-			"link": "https://kovan.etherscan.io/address/0x8F058332CdE72FEb4E062aAB521C5659A36E88b2",
-			"timestamp": "2020-02-17T13:45:25.799Z",
-			"txn": "",
+			"link": "https://kovan.etherscan.io/address/0x73a61f7D7401096ccB4AC883ebeb7bE7605c3955",
+			"timestamp": "2019-09-09T00:16:44.000Z",
+			"txn": "https://kovan.etherscan.io/tx/0xdafa2d664946ed01ed921bd9464066b7a15d681586bc0aabb98ff9ef648ec463",
 			"network": "kovan"
 		},
 		"ExchangeRates": {
@@ -3273,8 +3273,7 @@
 				{
 					"inputs": [
 						{
-							"indexed": true,
-							"name": "account",
+							"name": "_proxy",
 							"type": "address"
 						},
 						{
@@ -5168,8 +5167,7 @@
 					],
 					"payable": false,
 					"stateMutability": "nonpayable",
-					"type": "constructor",
-					"signature": "constructor"
+					"type": "constructor"
 				},
 				{
 					"anonymous": false,
@@ -6460,10 +6458,6 @@
 						{
 							"name": "spender",
 							"type": "address"
-						},
-						{
-							"name": "value",
-							"type": "uint256"
 						}
 					],
 					"name": "allowance",
@@ -6474,7 +6468,7 @@
 						}
 					],
 					"payable": false,
-					"stateMutability": "nonpayable",
+					"stateMutability": "view",
 					"type": "function",
 					"signature": "0xdd62ed3e"
 				},
@@ -6509,7 +6503,7 @@
 				{
 					"constant": true,
 					"inputs": [],
-					"name": "FEE_ADDRESS",
+					"name": "tokenState",
 					"outputs": [
 						{
 							"name": "",
@@ -6519,7 +6513,7 @@
 					"payable": false,
 					"stateMutability": "view",
 					"type": "function",
-					"signature": "0xeb1edd61"
+					"signature": "0xe90dd9e2"
 				},
 				{
 					"constant": true,
@@ -7657,8 +7651,7 @@
 					],
 					"payable": false,
 					"stateMutability": "nonpayable",
-					"type": "constructor",
-					"signature": "constructor"
+					"type": "constructor"
 				},
 				{
 					"payable": true,
